refactor: Dropdown 위치 이동
